Optimized search
- Reinstated top-level alpha and beta updates. This will allow for
faster cutoffs of the search tree at the expense of having to continue
searching if there is only one action possible that does not lead to a
loss. However, this additional search is now worth it since it utilizes
our time limit to add nodes to the transposition table.
- Cleared the transposition table when we are guaranteed to lose. This
will ensure that we recompute the best action each time, stalling out
the loss as much as possible.